+++ datacapture/src/main/java/com/google/android/fhir/datacapture/contrib/views/barcode/mlkit/md/barcodedetection/BarcodeProcessor.kt
@@ -14,18 +14,18 @@
  * limitations under the License.
  */
 
-package com.google.android.fhir.datacapture.views.barcode.mlkit.md.barcodedetection
+package com.google.android.fhir.datacapture.contrib.views.barcode.mlkit.md.barcodedetection
 
 import android.animation.ValueAnimator
 import android.util.Log
 import androidx.annotation.MainThread
-import com.google.android.fhir.datacapture.views.barcode.mlkit.md.InputInfo
-import com.google.android.fhir.datacapture.views.barcode.mlkit.md.camera.CameraReticleAnimator
-import com.google.android.fhir.datacapture.views.barcode.mlkit.md.camera.FrameProcessorBase
-import com.google.android.fhir.datacapture.views.barcode.mlkit.md.camera.GraphicOverlay
-import com.google.android.fhir.datacapture.views.barcode.mlkit.md.camera.WorkflowModel
-import com.google.android.fhir.datacapture.views.barcode.mlkit.md.camera.WorkflowModel.WorkflowState
-import com.google.android.fhir.datacapture.views.barcode.mlkit.md.settings.PreferenceUtils
+import com.google.android.fhir.datacapture.contrib.views.barcode.mlkit.md.InputInfo
+import com.google.android.fhir.datacapture.contrib.views.barcode.mlkit.md.camera.CameraReticleAnimator
+import com.google.android.fhir.datacapture.contrib.views.barcode.mlkit.md.camera.FrameProcessorBase
+import com.google.android.fhir.datacapture.contrib.views.barcode.mlkit.md.camera.GraphicOverlay
+import com.google.android.fhir.datacapture.contrib.views.barcode.mlkit.md.camera.WorkflowModel
+import com.google.android.fhir.datacapture.contrib.views.barcode.mlkit.md.camera.WorkflowModel.WorkflowState
+import com.google.android.fhir.datacapture.contrib.views.barcode.mlkit.md.settings.PreferenceUtils
 import com.google.android.gms.tasks.Task
 import com.google.mlkit.vision.barcode.Barcode
 import com.google.mlkit.vision.barcode.BarcodeScanning
