+++ datacapture/src/main/java/com/google/android/fhir/datacapture/contrib/views/barcode/mlkit/md/camera/FrameProcessorBase.kt
@@ -14,14 +14,14 @@
  * limitations under the License.
  */
 
-package com.google.android.fhir.datacapture.views.barcode.mlkit.md.camera
+package com.google.android.fhir.datacapture.contrib.views.barcode.mlkit.md.camera
 
 import android.os.SystemClock
 import android.util.Log
 import androidx.annotation.GuardedBy
-import com.google.android.fhir.datacapture.views.barcode.mlkit.md.CameraInputInfo
-import com.google.android.fhir.datacapture.views.barcode.mlkit.md.InputInfo
-import com.google.android.fhir.datacapture.views.barcode.mlkit.md.ScopedExecutor
+import com.google.android.fhir.datacapture.contrib.views.barcode.mlkit.md.CameraInputInfo
+import com.google.android.fhir.datacapture.contrib.views.barcode.mlkit.md.InputInfo
+import com.google.android.fhir.datacapture.contrib.views.barcode.mlkit.md.ScopedExecutor
 import com.google.android.gms.tasks.OnFailureListener
 import com.google.android.gms.tasks.Task
 import com.google.android.gms.tasks.TaskExecutors
