+++ datacapture/src/main/java/com/google/android/fhir/datacapture/contrib/views/barcode/mlkit/md/camera/WorkflowModel.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.google.android.fhir.datacapture.views.barcode.mlkit.md.camera
+package com.google.android.fhir.datacapture.contrib.views.barcode.mlkit.md.camera
 
 import android.app.Application
 import android.content.Context
